Refactoring. New files for extensions and constants.
