+++ PermissionScope/Constants.swift
@@ -1,0 +1,16 @@
+//
+//  Constants.swift
+//  PermissionScope
+//
+//  Created by Nick O'Neill on 8/21/15.
+//  Copyright © 2015 That Thing in Swift. All rights reserved.
+//
+
+import Foundation
+
+enum PermissionScopeConstants {
+    static let requestedInUseToAlwaysUpgrade = "requestedInUseToAlwaysUpgrade"
+    static let requestedForBluetooth = "askedForBluetooth"
+    static let requestedForMotion = "askedForMotion"
+    static let askedForNotificationsDefaultsKey = "PermissionScopeAskedForNotificationsDefaultsKey"
+}
