+++ PermissionScope/Extensions.swift
@@ -1,0 +1,25 @@
+//
+//  Extensions.swift
+//  PermissionScope
+//
+//  Created by Nick O'Neill on 8/21/15.
+//  Copyright © 2015 That Thing in Swift. All rights reserved.
+//
+
+import Foundation
+
+extension UIColor {
+    var inverseColor: UIColor{
+        var r:CGFloat = 0.0; var g:CGFloat = 0.0; var b:CGFloat = 0.0; var a:CGFloat = 0.0;
+        if self.getRed(&r, green: &g, blue: &b, alpha: &a) {
+            return UIColor(red: 1.0-r, green: 1.0 - g, blue: 1.0 - b, alpha: a)
+        }
+        return self
+    }
+}
+
+extension String {
+    var localized: String {
+        return NSLocalizedString(self, comment: "")
+    }
+}
