+++ PermissionScope/PermissionScope.swift
@@ -14,15 +14,10 @@ import Photos
 import EventKit
 import CoreBluetooth
 import CoreMotion
-
-struct PermissionScopeConstants {
-    static let requestedInUseToAlwaysUpgrade = "requestedInUseToAlwaysUpgrade"
-    static let requestedForBluetooth = "askedForBluetooth"
-    static let requestedForMotion = "askedForMotion"
-}
+import HealthKit
 
 @objc public enum PermissionType: Int, CustomStringConvertible {
-    case Contacts, LocationAlways, LocationInUse, Notifications, Microphone, Camera, Photos, Reminders, Events, Bluetooth, Motion
+    case Contacts, LocationAlways, LocationInUse, Notifications, Microphone, Camera, Photos, Reminders, Events, Bluetooth, Motion, HealthKit
     
     public var description: String {
         switch self {
@@ -37,6 +32,7 @@ struct PermissionScopeConstants {
         case .Reminders: return "Reminders"
         case .Bluetooth: return "Bluetooth"
         case .Motion: return "Motion"
+        case .HealthKit: return "HealthKit"
         }
     }
     
@@ -49,7 +45,7 @@ struct PermissionScopeConstants {
         }
     }
     
-    static let allValues = [Contacts, LocationAlways, LocationInUse, Notifications, Microphone, Camera, Photos, Reminders, Events, Bluetooth, Motion]
+    static let allValues = [Contacts, LocationAlways, LocationInUse, Notifications, Microphone, Camera, Photos, Reminders, Events, Bluetooth, Motion, HealthKit]
     
 }
 
@@ -77,8 +73,6 @@ struct PermissionScopeConstants {
         }
     }
 }
-
-private let PermissionScopeAskedForNotificationsDefaultsKey = "PermissionScopeAskedForNotificationsDefaultsKey"
 
 @objc public class PermissionConfig: NSObject {
     let type: PermissionType
@@ -115,26 +109,11 @@ private let PermissionScopeAskedForNotificationsDefaultsKey = "PermissionScopeAs
     }
 }
 
-extension UIColor {
-    var inverseColor: UIColor{
-        var r:CGFloat = 0.0; var g:CGFloat = 0.0; var b:CGFloat = 0.0; var a:CGFloat = 0.0;
-        if self.getRed(&r, green: &g, blue: &b, alpha: &a) {
-            return UIColor(red: 1.0-r, green: 1.0 - g, blue: 1.0 - b, alpha: a)
-        }
-        return self
-    }
-}
-extension String {
-    var localized: String {
-        return NSLocalizedString(self, comment: "")
-    }
-}
-
 @objc public class PermissionScope: UIViewController, CLLocationManagerDelegate, UIGestureRecognizerDelegate, CBPeripheralManagerDelegate {
     // constants
     let contentWidth: CGFloat = 280.0
 
-    // configurable things
+    // MARK: UI Parameters
     public let headerLabel = UILabel(frame: CGRect(x: 0, y: 0, width: 200, height: 50))
     public let bodyLabel = UILabel(frame: CGRect(x: 0, y: 0, width: 240, height: 70))
     public var tintColor = UIColor(red: 0, green: 0.47, blue: 1, alpha: 1)
@@ -143,7 +122,7 @@ extension String {
     public var closeButton = UIButton(frame: CGRect(x: 0, y: 0, width: 50, height: 32))
     public var closeOffset = CGSize(width: 0, height: 0)
 
-    // some view hierarchy
+    // MARK: View hierarchy for custom alert
     let baseView = UIView()
     let contentView = UIView()
 
@@ -162,18 +141,23 @@ extension String {
         return CMMotionActivityManager()
     }()
     
+    // Default status for CoreMotion
     var motionPermissionStatus: PermissionStatus = .Unknown
 
-    // internal state and resolution
+    // Internal state and resolution
     var configuredPermissions: [PermissionConfig] = []
     var permissionButtons: [UIButton] = []
     var permissionLabels: [UILabel] = []
 	
-	// properties that may be useful for direct use of the request* methods
-    public var authChangeClosure: ((Bool, [PermissionResult]) -> Void)? = nil
-    public var cancelClosure: (([PermissionResult]) -> Void)? = nil
-	/** Called when the user has disabled or denied access to notifications, and we're presenting them with a help dialog. */
-    public var disabledOrDeniedClosure: (([PermissionResult]) -> Void)? = nil
+	// Useful for direct use of the request* methods
+    public var authChangeClosure: authClosureType? = nil
+    public typealias authClosureType = (Bool, [PermissionResult]) -> Void
+    
+    public var cancelClosure: cancelClosureType? = nil
+	public typealias cancelClosureType = ([PermissionResult]) -> Void
+    
+    /** Called when the user has disabled or denied access to notifications, and we're presenting them with a help dialog. */
+    public var disabledOrDeniedClosure: cancelClosureType? = nil
 	/** View controller to be used when presenting alerts. Defaults to self. You'll want to set this if you are calling the `request*` methods directly. */
 	public var viewControllerForAlerts : UIViewController?
 
@@ -341,7 +325,7 @@ extension String {
         }
     }
 
-    // MARK: customizing the permissions
+    // MARK: - Customizing the permissions
 
     @objc public func addPermission(config: PermissionConfig) {
         assert(!config.message.isEmpty, "Including a message about your permission usage is helpful")
@@ -405,8 +389,9 @@ extension String {
         return label
     }
 
-    // MARK: status and requests for each permission
-
+    // MARK: - Status and Requests for each permission
+    
+    // MARK: Location
     public func statusLocationAlways() -> PermissionStatus {
         if !CLLocationManager.locationServicesEnabled() {
             return .Disabled
@@ -481,6 +466,7 @@ extension String {
         }
     }
 
+    // MARK: Contacts
     public func statusContacts() -> PermissionStatus {
         let status = ABAddressBookGetAuthorizationStatus()
         switch status {
@@ -506,13 +492,13 @@ extension String {
         }
     }
 
-    // TODO: can we tell if notifications has been denied?
+    // MARK: Notifications
     public func statusNotifications() -> PermissionStatus {
         let settings = UIApplication.sharedApplication().currentUserNotificationSettings()
         if let settingTypes = settings?.types where settingTypes != UIUserNotificationType.None {
             return .Authorized
         } else {
-            if NSUserDefaults.standardUserDefaults().boolForKey(PermissionScopeAskedForNotificationsDefaultsKey) {
+            if NSUserDefaults.standardUserDefaults().boolForKey(PermissionScopeConstants.askedForNotificationsDefaultsKey) {
                 return .Unauthorized
             } else {
                 return .Unknown
@@ -555,7 +541,7 @@ extension String {
             // There should be only one...
             let notificationsPermissionSet = self.configuredPermissions.filter { $0.notificationCategories != .None && !$0.notificationCategories!.isEmpty }.first?.notificationCategories
             
-            NSUserDefaults.standardUserDefaults().setBool(true, forKey: PermissionScopeAskedForNotificationsDefaultsKey)
+            NSUserDefaults.standardUserDefaults().setBool(true, forKey: PermissionScopeConstants.askedForNotificationsDefaultsKey)
             NSUserDefaults.standardUserDefaults().synchronize()
             
             NSNotificationCenter.defaultCenter().addObserver(self, selector: Selector("showingNotificationPermission"), name: UIApplicationWillResignActiveNotification, object: nil)
@@ -574,6 +560,7 @@ extension String {
         }
     }
     
+    // MARK: Microphone
     public func statusMicrophone() -> PermissionStatus {
         let status = AVAudioSession.sharedInstance().recordPermission()
         if status == .Granted {
@@ -598,6 +585,7 @@ extension String {
         }
     }
     
+    // MARK: Camera
     public func statusCamera() -> PermissionStatus {
         let status = AVCaptureDevice.authorizationStatusForMediaType(AVMediaTypeVideo)
         switch status {
@@ -624,6 +612,7 @@ extension String {
         }
     }
 
+    // MARK: Photos
     public func statusPhotos() -> PermissionStatus {
         let status = PHPhotoLibrary.authorizationStatus()
         switch status {
@@ -649,6 +638,7 @@ extension String {
         }
     }
     
+    // MARK: Reminders
     public func statusReminders() -> PermissionStatus {
         let status = EKEventStore.authorizationStatusForEntityType(EKEntityType.Reminder)
         switch status {
@@ -675,6 +665,7 @@ extension String {
         }
     }
     
+    // MARK: Events
     public func statusEvents() -> PermissionStatus {
         let status = EKEventStore.authorizationStatusForEntityType(EKEntityType.Event)
         switch status {
@@ -701,6 +692,7 @@ extension String {
         }
     }
     
+    // MARK: Bluetooth
     private var askedBluetooth:Bool {
         get {
             return NSUserDefaults.standardUserDefaults().boolForKey(PermissionScopeConstants.requestedForBluetooth)
@@ -759,18 +751,7 @@ extension String {
         }
     }
     
-    private var askedMotion:Bool {
-        get {
-            return NSUserDefaults.standardUserDefaults().boolForKey(PermissionScopeConstants.requestedForMotion)
-        }
-        set {
-            NSUserDefaults.standardUserDefaults().setBool(newValue, forKey: PermissionScopeConstants.requestedForMotion)
-            NSUserDefaults.standardUserDefaults().synchronize()
-        }
-    }
-    
-    private var waitingForMotion = false
-    
+    // MARK: CoreMotion
     public func statusMotion() -> PermissionStatus {
         if askedMotion{
             triggerMotionStatusUpdate()
@@ -812,7 +793,20 @@ extension String {
         waitingForMotion = true
     }
     
-    public func statusHealth() -> PermissionStatus {
+    private var askedMotion:Bool {
+        get {
+            return NSUserDefaults.standardUserDefaults().boolForKey(PermissionScopeConstants.requestedForMotion)
+        }
+        set {
+            NSUserDefaults.standardUserDefaults().setBool(newValue, forKey: PermissionScopeConstants.requestedForMotion)
+            NSUserDefaults.standardUserDefaults().synchronize()
+        }
+    }
+    
+    private var waitingForMotion = false
+    
+    // MARK: HealthKit
+    public func statusHealthKit() -> PermissionStatus {
         // There should be only one...
         // TODO: Refactor to comp. var.
         //        let healthCategoryTypes = self.configuredPermissions.filter { $0.healthCategoryTypes != .None && !$0.healthCategoryTypes!.isEmpty }.first?.healthCategoryTypes
@@ -828,7 +822,7 @@ extension String {
         return .Unknown
     }
     
-    func requestHealth() {
+    func requestHealthKit() {
         //        switch statusHealth() {
         //        case .Unknown:
         //            HKHealthStore().requestAuthorizationToShareTypes(<#typesToShare: Set<NSObject>!#>,
@@ -845,7 +839,7 @@ extension String {
         //        }
     }
     
-    // MARK: finally, displaying the panel
+    // MARK: - UI
 
     @objc public func show(authChange: ((finished: Bool, results: [PermissionResult]) -> Void)? = nil, cancelled: ((results: [PermissionResult]) -> Void)? = nil) {
         assert(configuredPermissions.count > 0, "Please add at least one permission")
@@ -926,52 +920,53 @@ extension String {
         notificationTimer = nil
     }
 
+    // MARK: - Gesture delegate
+
+    public func gestureRecognizer(gestureRecognizer: UIGestureRecognizer, shouldReceiveTouch touch: UITouch) -> Bool {
+
+        // this prevents our tap gesture from firing for subviews of baseview
+        if touch.view == baseView {
+            return true
+        }
+
+        return false
+    }
+
+    // MARK: - Location delegate
+
+    public func locationManager(manager: CLLocationManager, didChangeAuthorizationStatus status: CLAuthorizationStatus) {
+
+        detectAndCallback()
+    }
+    
+    // MARK: - Bluetooth delegate
+    
+    public func peripheralManagerDidUpdateState(peripheral: CBPeripheralManager) {
+        
+        waitingForBluetooth = false
+        detectAndCallback()
+    }
+
+    // MARK: - UI Helpers
+    
     func cancel() {
         self.hide()
-
+        
         if let cancelClosure = cancelClosure {
             cancelClosure(getResultsForConfig())
         }
     }
-
+    
     func finish() {
         self.hide()
-
+        
         if let authChangeClosure = authChangeClosure {
             authChangeClosure(true, getResultsForConfig())
         }
     }
-
-    func detectAndCallback() {
-        // compile the results and pass them back if necessary
-        if let authChangeClosure = authChangeClosure {
-            authChangeClosure(allAuthorized, getResultsForConfig())
-        }
-
-        dispatch_async(dispatch_get_main_queue(), { () -> Void in
-            self.view.setNeedsLayout()
-
-            // and hide if we've sucessfully got all permissions
-            if self.allAuthorized {
-                self.hide()
-            }
-        })
-    }
-
-    func getResultsForConfig() -> [PermissionResult] {
-        var results: [PermissionResult] = []
-
-        for config in configuredPermissions {
-            let status = statusForPermission(config.type)
-            let result = PermissionResult(type: config.type, status: status, demands: config.demands)
-            results.append(result)
-        }
-
-        return results
-    }
-
+    
     func appForegroundedAfterSettings (){
-
+        
         NSNotificationCenter.defaultCenter().removeObserver(self, name: UIApplicationDidBecomeActiveNotification, object: nil)
         
         detectAndCallback()
@@ -1012,33 +1007,6 @@ extension String {
         viewControllerForAlerts?.presentViewController(alert,
             animated: true, completion: nil)
     }
-    
-    // MARK: gesture delegate
-
-    public func gestureRecognizer(gestureRecognizer: UIGestureRecognizer, shouldReceiveTouch touch: UITouch) -> Bool {
-
-        // this prevents our tap gesture from firing for subviews of baseview
-        if touch.view == baseView {
-            return true
-        }
-
-        return false
-    }
-
-    // MARK: location delegate
-
-    public func locationManager(manager: CLLocationManager, didChangeAuthorizationStatus status: CLAuthorizationStatus) {
-
-        detectAndCallback()
-    }
-    
-    // MARK: bluetooth delegate
-    
-    public func peripheralManagerDidUpdateState(peripheral: CBPeripheralManager) {
-        
-        waitingForBluetooth = false
-        detectAndCallback()
-    }
 
     // MARK: Helpers
     
@@ -1067,6 +1035,36 @@ extension String {
             return statusBluetooth()
         case .Motion:
             return statusMotion()
+        case .HealthKit:
+            return statusHealthKit()
         }
+    }
+    
+    func detectAndCallback() {
+        // compile the results and pass them back if necessary
+        if let authChangeClosure = authChangeClosure {
+            authChangeClosure(allAuthorized, getResultsForConfig())
+        }
+        
+        dispatch_async(dispatch_get_main_queue(), { () -> Void in
+            self.view.setNeedsLayout()
+            
+            // and hide if we've sucessfully got all permissions
+            if self.allAuthorized {
+                self.hide()
+            }
+        })
+    }
+    
+    func getResultsForConfig() -> [PermissionResult] {
+        var results: [PermissionResult] = []
+        
+        for config in configuredPermissions {
+            let status = statusForPermission(config.type)
+            let result = PermissionResult(type: config.type, status: status, demands: config.demands)
+            results.append(result)
+        }
+        
+        return results
     }
 }
